CC-2894 Escape related entity embedded content
The view page implements links to related entities by embedding some
of the descriptive elements (title, description, etc.) within HTML
attributes.

The content was not properly escaped, so that if the content contained
an apostrophe, this terminated the HTML attribute. If the apostrophe
was in the title, this could break the display of the title. But in
any case, it broke the link to bring up the related entity modal.

Add escaping of the content.

Note the explicit specification of the second parameter, i.e.,
"htmlspecialchars(..., ENT_QUOTES)".  As the Changelog at
https://www.php.net/manual/en/function.htmlspecialchars.php notes, the
default value only changed to include ENT_QUOTES in PHP 8.1.0; we're
still using PHP 5.

RVA-241