CC-2919 Percent-encode auth header for Registry
Oops, when preparing the authentication header for use with the
Registry API, we were sending "raw" data, i.e., that could contain
non-ASCII characters.

The contents of the header are the session cookie that CodeIgniter
creates. CodeIgniter percent-encodes the contents when sending it as a
cookie. But we get that data by asking CodeIgniter, and CodeIgniter
gives it back _without_ the percent encoding.

The Registry expects the authentication header to be percent-encoded.

So, percent encode the contents of the authentication header.

RVA-252