CC-2921 Fix related vocab "View vocabulary" button
If a vocabulary has a related internal vocabulary, that is indicated
on the vocabulary's view page in a right-hand "Related" panel.

If you click the link for the related vocabulary, you get a popup that
includes a button with the label "View vocabulary".

The button is not implemented correctly. If you move the mouse pointer
down the popup so that it passes just over the top part of the button,
the button's foreground and background colours swap to indicate that
the button can be selected. But if you click, nothing happens. You
have to move the mouse pointer down further, over the words "View
vocabulary", at which point you'll see those words underlined, and
then you can click to bring up the related vocabulary.

Fix the button so that it works like the other buttons in the
interface, i.e., so that all of the button is active, and remove the
underlining behaviour.

RVA-253