CC-2179 Show Registry contraint violations better
If the Portal sent a PUT on the vocab, and the Registry returned
constraint violations, and if there was more than one violation with
the same message text, none of the violations appeared in the Portal.

Example:

$scope.errors is set to:

["Subject IRIs may not be duplicated",
 "Vocabulary subject IRI not recognised",
 "Vocabulary subject IRI not recognised"]

It's perfectly possible/normal for the same error to occur multiple
times, if the error is for a repeated element (in this case,
subjects).

In this case, the div with ng-repeat over errors fails with an
"ngRepeat:dupes" error.

See
https://code.angularjs.org/1.6.6/docs/error/ngRepeat/dupes?p0=error%20in%20errors&p1=string:Vocabulary%20subject%20IRI%20not%20recognised&p2=Vocabulary%20subject%20IRI%20not%20recognised

But, in fact, it would be "good" to show the location details included
with the validation errors. In the above example, the two instances of
the "Vocabulary subject IRI not recognised" come with an indication of
which subject's IRI was not recognised.

So this is what has been done:

1. Don't fail if the there's a duplicate error message. See
   cms.blade.php for the change, done exactly as per the advice as
   code.angularjs.org.

2. Append path information to the error message, where it's
   available. See vocabs_cms.js for the change.

It turns out that the work for 2. then "overrides" the work for 1. in
the specific example that prompted this change ... but "never mind"
... it may come up again in another context.

RVA-169