Update package-lock.json without hash
See commit f0c0d063 for the failed first attempt to do this.

Needed to run "npm install" due to an update to the Vocabs Registry's
enumerated types.

This updated package-log.json; but there seems to be no other
substantive difference to the contents of node_modules.

Do again what I did in commit f0c0d063, but this time, follow my own
instructions to post-process to remove the hash for the Registry API
client library.