CC-2920 Wait for logCms to complete before navigating away
When a user clicks one of the save-type action buttons on the CMS
(i.e., "Save to draft", "Publish", or "Deprecate"), the action is
performed by (client-side) JavaScript code invoking methods of the
Registry API, i.e., without invoking a Portal-side PHP controller
method. If the action succeeds, the JavaScript code modifies
window.location in order to perform a browser-side redirect (either to
the view page, or (in the case of "Save to draft") back to itself).

However, we want to have some Portal-side logging of the action, so
the JavaScript does in fact invoke a PHP controller method ... i.e.,
logCms, specifically just for logging. This is done using $.ajax().

This "normally" seems to work fine. However, while using Firefox to
investigate another issue, I've noticed that the browser (i.e.,
Firefox, not the user directly) sometimes cancels this AJAX request
before the request is sent to the server. This can be observed in the
Firefox Developer Tools Network tab, where the request is shown with
an error icon (or a blank) in the Status column, and
NS_BINDING_ABORTED (or 0 B) in the Transferred column. (See attached
screenshot.) And indeed, there's no log entry for the request (in the
Apache HTTP Server logs).

It's a defect in the (i.e., my) code: specifically, a race condition.

In order to make sure that the browser always sends the logCms AJAX
request to the server, have the CMS JavaScript code store the return
value of $.ajax(), which is a Promise-like value, and then only do the
redirect after the promise has completed.

RVA-291