CC-2723 Improve validation of RE "web pages" and URIs
This is continuation of a general improvement of the validation of
URLs and URIs.

Apply the new validation used for URLs given as access points also for
"web pages" of related entities.

For now, leave validation of URI identifiers as-is, but add a non-URL
URI example to the placeholder text.

RVA-296