CC-2539 Change widget default endpoint to correct value
The location of the proxy for the vocabulary widget was changed a long
time ago, but we didn't update the default value for the "endpoint"
parameter to match.

Change the default value for the "endpoint" parameter to match the
correct location.

Take the opportunity to do yet another instance of rebranding ANDS ->
ARDC.

RVADEV-6