CC-1505 Allow a vocabulary's owner to be changed
Once the CMS's vocabulary Owner field was set, the field stayed set
with that value.

The Registry now supports changing a vocabulary's owner; allow Portal
users to make use of that functionality.

The Owner dropdown now allows its value to be changed.

NB: despite this change, we do still maintain the multi-step process
of vocabulary creation where the user must specify a vocabulary's
owner _before_ being able to specify other metadata values. This is
because an owner must be known before related entities are added; an
owner must be specified for them when they're created. This is
particularly important when adding a new PoolParty project, where
pre-filling the form typically adds new related entities.

RVADEV-11